\name{c.months}
\alias{c.months}
\alias{c.mon}
\alias{c.weekdays}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ Names of months and weekdays }
\description{Vector(s) containing the months' and weekdays' names.
c.mon contains the 3 letter abbreviations of the months}
\usage{
c.months
c.mon
c.weekdays
}
%- maybe also 'usage' for other objects documented here.
\value{
  vector of names
}
\author{Werner A. Stahel, ETH Zurich}
\examples{
c.months
c.mon[4:6]
c.weekdays
}
\keyword{miscellaneous}
