\name{regr0-internal}
\title{Internal regr0 objects}
\alias{.Options}
\alias{BR}
\alias{DB}
\alias{add1.default}
\alias{deviance.lm}
\alias{drop1.default}
\alias{drop1.lm}
\alias{extractAIC.regr}
\alias{fitted.values.polr}
\alias{i.add1na}
\alias{i.glm}
\alias{i.lm}
\alias{i.main}
\alias{i.mlm}
\alias{i.mlmsum}
\alias{i.multinomial}
\alias{i.survreg}
\alias{i.plotlws}
\alias{i.plotlars}
\alias{i.polr}
\alias{is.formula}
\alias{i.smooth}
\alias{i.termtable}
\alias{is.R}
\alias{merge1}
\alias{nna}
\alias{nafalse}
\alias{print.mregr}
\alias{summary.mreg}
\alias{summary.regr}
\alias{u.merge}
\alias{vif.regr}
\alias{IR}
% \alias{"%nin%"}
\description{Internal regr0 objects.}
\details{These are not to be called by the user.}
\keyword{internal}
